(bug) update rpm to float in all places (#87)
Looks like af7501b missed a few places, which results
in A4988 and DRV8880 truncating an eventual float rpm.

+++ src/A4988.h
@@ -45,7 +45,7 @@ public:
     A4988(short steps, short dir_pin, short step_pin);
     A4988(short steps, short dir_pin, short step_pin, short enable_pin);
     
-    void begin(short rpm=60, short microsteps=1);
+    void begin(float rpm=60, short microsteps=1);
     /*
      * Fully wired. All the necessary control pins for A4988 are connected.
      */
